% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{as.data.frame.tc}
\alias{as.data.frame.tc}
\title{Cast a \code{tc} object to a data frame}
\usage{
\method{as.data.frame}{tc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tc}.}

\item{...}{Additional arguments; ignored.}
}
\value{
A data frame with a column \code{time} with the values in \code{x}
}
\description{
Cast a \code{tc} object to a data frame
}
