% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{c.tc}
\alias{c.tc}
\title{Combining \code{tc} objects}
\usage{
\method{c}{tc}(..., recursive = FALSE)
}
\arguments{
\item{...}{One or multiple \code{tc} objects to combine.}

\item{recursive}{A logical value; ignored.}
}
\description{
Combining \code{tc} objects
}
