% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{range.tc}
\alias{range.tc}
\title{Range of \code{tc} objects.}
\usage{
\method{range}{tc}(..., na.rm = FALSE)
}
\arguments{
\item{...}{\code{tc} objects.}

\item{na.rm}{A logical value; Indicating if \code{NA}s should be omitted.}
}
\description{
Range of \code{tc} objects.
}
