% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftime.R
\name{st_get_time}
\alias{st_get_time}
\title{Extract the active time column of an \code{sftime} object.}
\usage{
st_get_time(x)
}
\arguments{
\item{x}{An \code{sftime} object.}
}
\value{
The active time column in \code{x} as \code{tc} object.
}
\description{
Extract the active time column of an \code{sftime} object.
}
