% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftime.R
\name{st_sf_time}
\alias{st_sf_time}
\title{combine sf and tc objects to an sftime object}
\usage{
st_sf_time(sf, time)
}
\arguments{
\item{sf}{an sf object}

\item{time}{an tc object, see \link{st_tc} for details}
}
\description{
combine sf and tc objects to an sftime object
}
