% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftime.R
\name{st_sftime}
\alias{st_sftime}
\title{Construct a \code{sftime} object from all its components}
\usage{
st_sftime(
  ...,
  agr = NA_agr_,
  row.names,
  stringsAsFactors = TRUE,
  crs,
  precision,
  sf_column_name = NULL,
  check_ring_dir = FALSE,
  sfc_last = TRUE
)
}
\arguments{
\item{...}{column elements to be binded into an \code{sftime} object or a single \code{list} or \code{data.frame} with such columns; at least one of these columns shall be a geometry list-column of class \code{sfc} and one shall be a temporal colum of class \code{tc}.}

\item{agr}{character vector; see details below.}

\item{row.names}{row.names for the created \code{sf} object}

\item{stringsAsFactors}{logical; see \link{st_read}}

\item{crs}{coordinate reference system, something suitable as input to \link{st_crs}}

\item{precision}{numeric; see \link{st_as_binary}}

\item{sf_column_name}{character; name of the active list-column with simple feature geometries; in case
there is more than one and \code{sf_column_name} is \code{NULL}, the first one is taken.}

\item{check_ring_dir}{see \link{st_read}}

\item{sfc_last}{logical; if \code{TRUE}, \code{sfc} columns are always put last, otherwise column order is left unmodified.}
}
\value{
an object of class \code{sftime}
}
\description{
Construct a \code{sftime} object from all its components
}
