% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{st_tc-extract}
\alias{st_tc-extract}
\alias{[.tc}
\alias{[<-.tc}
\title{Extract or replace elements in a \code{tc} object}
\usage{
\method{[}{tc}(x, i, ...)

\method{[}{tc}(x, i) <- value
}
\arguments{
\item{x}{The \code{\link[=st_tc]{tc}} object from which to extract 
element(s) or in which to replace elements.}

\item{i}{Any subsetting expression supported by the temporal class provided 
by the \code{tc} object.}

\item{...}{any further arguments for the underlying subsetting method.}

\item{value}{An object with which to substitute elements in \code{x}. Must 
be of the same class as elements in \code{x}.}
}
\value{
A \code{tc} object representing a subset of \code{x} or \code{x} 
with replaced values.
}
\description{
Extract or replace elements in a \code{tc} object
}
