% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{st_tc}
\alias{st_tc}
\title{Construct a \code{tc} object from a vector with time information}
\usage{
st_tc(x)
}
\arguments{
\item{x}{A vector or list.}
}
\value{
An object of class \code{tc}.
}
\description{
An object of class \code{tc} stores time information and can be used as
time column in a \code{\link[=st_sftime]{sftime}} object.
}
