% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{str.tc}
\alias{str.tc}
\title{Compactly display the structure of \code{tc} objects}
\usage{
\method{str}{tc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{tc}.}

\item{...}{Additional arguments passed to the \code{str} method for the 
first element in \code{object}.}
}
\description{
Compactly display the structure of \code{tc} objects
}
