% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{[.tc}
\alias{[.tc}
\title{Subsetting of time columns \code{tc}}
\usage{
\method{[}{tc}(x, i, ...)
}
\arguments{
\item{x}{the object to be subsetted}

\item{i}{any subsetting expression supported by the temporal class provided to the \code{tc}.}

\item{...}{any further arguments for the underlying subsetting function}
}
\value{
a time column \code{tc} holding a subset of the orignal time column
}
\description{
Subsetting of time columns \code{tc}
}
