% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{summary.tc}
\alias{summary.tc}
\title{Summarize \code{tc} objects}
\usage{
\method{summary}{tc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{tc}.}

\item{...}{Additional arguments; Ignored.}
}
\description{
Summarize \code{tc} objects
}
