% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{timeColumnIsSortable}
\alias{timeColumnIsSortable}
\title{Check the time column}
\usage{
timeColumnIsSortable(x)
}
\arguments{
\item{x}{the object to check}
}
\description{
Checks whether the time column is sortable.
}
\details{
checks whether the provided object can be handled by \link{order}. A couple of basic types are whitelisted. However, custom types can be defined when they provide a dedicated generic to \link{xtfrm}. Note that a \code{list} can only be sorted with \link{atomic} values. See the examples below for a template.
}
