% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sftime}
\alias{plot.sftime}
\alias{plot}
\title{Plot an sftime object}
\usage{
\method{plot}{sftime}(x, y, ..., mode = "xy", number = 6, tcuts, key.pos = 0)
}
\arguments{
\item{x}{the sftime object to be plotted}

\item{y}{the variabel name to be plotted; if missing, the first variable is plotted}

\item{...}{passed on to nested plot functions}

\item{mode}{how to project the 3D object to a 2D device: 'xy' = separate maps for time ranges}

\item{number}{of panels to be plotted, cannot be larger than the number of timestamps; ignored when 'tcuts' is provided}

\item{tcuts}{predefined temporal ranges assigned to each map; if missing, will be determined as equal spans according to 'number'}

\item{key.pos}{see \link[sf]{plot_sf}}
}
\description{
Plot an sftime object
}
\examples{
set.seed(123)
coords <- matrix(runif(100), ncol = 2)
g = st_sfc(lapply(1:50, function(i) st_point(coords[i,]) ))
sf <- st_sf(a=1:50, g)

sft <- st_sf_time(sf, st_tc(as.POSIXct("2020-09-01 00:00:00")+0:49*3600*6))
plot(sft)

}
