% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftime.R
\name{print.sftime}
\alias{print.sftime}
\title{Print a \code{sftime} object}
\usage{
\method{print}{sftime}(x, ..., n = getOption("sf_max_print", default = 10))
}
\arguments{
\item{x}{the object to be printed}

\item{...}{currently unused, for compatability}

\item{n}{maximum number of printed elements}
}
\value{
a human readble print out on the console
}
\description{
Print a \code{sftime} object
}
