% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftime.R
\name{print.sftime}
\alias{print.sftime}
\title{Print a \code{sftime} object}
\usage{
\method{print}{sftime}(x, ..., n = getOption("sf_max_print", default = 10))
}
\arguments{
\item{x}{An object of class \code{sftime}.}

\item{...}{Currently unused arguments, for compatibility.}

\item{n}{Numeric value; maximum number of printed elements.}
}
\value{
\code{x} (invisible).
}
\description{
Print a \code{sftime} object
}
