% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{print.tc}
\alias{print.tc}
\title{\code{print} generic for \code{tc}}
\usage{
\method{print}{tc}(x, ..., noFeature = FALSE)
}
\arguments{
\item{x}{an object of class \code{tc}}

\item{...}{currently unused, for compatability}

\item{noFeature}{logical: whether the number of features shall not be printed (default: FALSE)}
}
\value{
human readable print on the console
}
\description{
\code{print} generic for \code{tc}
}
