% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftime.R
\name{st_as_sftime}
\alias{st_as_sftime}
\alias{st_as_sftime.ST}
\title{Convert foreign object to an sftime object}
\usage{
st_as_sftime(x)
}
\arguments{
\item{x}{object to be converted into an object class \code{sftime}}
}
\description{
Convert foreign object to an sftime object
}
