% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_time.R
\name{st_time}
\alias{st_time}
\alias{st_time.tc}
\title{Get, set, or replace time information}
\usage{
st_time(obj, ...)

\method{st_time}{tc}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{sftime} or \code{tc}.}

\item{...}{Additional arguments; Ignored.}
}
\value{
An object of class \code{\link[=st_tc]{tc}}
}
\description{
Get, set, or replace time information
}
