% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.R
\name{[.tc}
\alias{[.tc}
\title{Subsetting of \code{tc} objects}
\usage{
\method{[}{tc}(x, i, ...)
}
\arguments{
\item{x}{The \code{\link{tc}} object to be subsetted.}

\item{i}{Any subsetting expression supported by the temporal class provided by the \code{tc} object.}

\item{...}{any further arguments for the underlying subsetting method.}
}
\value{
A \code{tc} object representing a subset of \code{x}.
}
\description{
Subsetting of \code{tc} objects
}
