% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftime.R
\name{transform.sftime}
\alias{transform.sftime}
\title{Transform method for \code{sftime} objects}
\usage{
\method{transform}{sftime}(`_data`, ...)
}
\arguments{
\item{_data}{An object of class \code{\link[=st_sftime]{sftime}}.}

\item{...}{Further arguments of the form new_variable=expression}
}
\description{
Can be used to create or modify attribute variables; for transforming 
geometries see \code{\link[sf]{st_transform}}, and all other functions starting with 
\code{st_}.
}
